Move hiredis back into the lib folder to keep it deprecated
