+++ lib/parser.js
@@ -254,7 +254,7 @@ function JavascriptRedisParser (options) {
   if (options.name === 'hiredis') {
     /* istanbul ignore next: hiredis is only supported for legacy usage */
     try {
-      var Hiredis = require('../test/hiredis')
+      var Hiredis = require('./hiredis')
       console.error(new TypeError('Using hiredis is discouraged. Please use the faster JS parser by removing the name option.').stack.replace('Error', 'Warning'))
       return new Hiredis(options)
     } catch (e) {
